pyproject.toml
[project]
name = "online-course-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "psycopg[binary]>=3.2.9",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
